[sub/estimate-1] 미요사 관련 모듈 코드 리팩토링 완료
