+++ daengle-groomer-api/src/main/java/ddog/groomer/application/GroomerRepository.java
@@ -1,10 +1,8 @@
 package ddog.groomer.application;
 
-import ddog.daengleserver.domain.account.Groomer;
+import ddog.domain.groomer.Groomer;
 
 public interface GroomerRepository {
-
-    String getAddressById(Long accountId);
 
     Groomer getGroomerByAccountId(Long accountId);
 }
