+++ daengle-groomer-api/src/main/java/ddog/groomer/application/GroomingEstimateRepository.java
@@ -1,6 +1,6 @@
 package ddog.groomer.application;
 
-import ddog.daengleserver.domain.estimate.GroomingEstimate;
+import ddog.domain.estimate.GroomingEstimate;
 
 import java.util.List;
 
@@ -9,8 +9,6 @@ public interface GroomingEstimateRepository {
     void save(GroomingEstimate groomingEstimate);
 
     GroomingEstimate getByGroomingEstimateId(Long groomingEstimateId);
-
-    List<GroomingEstimate> findGroomingEstimatesByPetId(Long petId);
 
     List<GroomingEstimate> findGeneralGroomingEstimates(String address);
