+++ daengle-vet-api/src/main/java/ddog/vet/application/VetRepository.java
@@ -4,8 +4,6 @@ import ddog.domain.vet.Vet;
 
 public interface VetRepository {
 
-    String getAddressById(Long accountId);
-
     Vet getVetByAccountId(Long accountId);
 
 }
